# perf(serde_v8): introduce Serializable boxable object