--- a/serde_v8/src/lib.rs
+++ b/serde_v8/src/lib.rs
@@ -5,6 +5,7 @@ mod keys;
 mod magic;
 mod payload;
 mod ser;
+mod serializable;
 pub mod utils;
 
 pub use de::{from_v8, from_v8_cached, Deserializer};
@@ -12,3 +13,4 @@ pub use error::{Error, Result};
 pub use keys::KeyCache;
 pub use magic::Value;
 pub use ser::{to_v8, Serializer};
+pub use serializable::Serializable;