--- a/serde_v8/src/serializable.rs
+++ b/serde_v8/src/serializable.rs
@@ -1,22 +1,22 @@
 use rusty_v8 as v8;
 
-/// V8Serializable exists to allow boxing values as "objects" to be serialized later,
+/// Serializable exists to allow boxing values as "objects" to be serialized later,
 /// this is particularly useful for async op-responses. This trait is a more efficient
 /// replacement for erased-serde that makes less allocations, since it's specific to serde_v8
 /// (and thus doesn't have to have generic outputs, etc...)
-pub trait V8Serializable {
+pub trait Serializable {
   fn to_v8<'a>(
     &self,
     scope: &mut v8::HandleScope<'a>,
-  ) -> Result<v8::Local<'a, v8::Value>, serde_v8::Error>;
+  ) -> Result<v8::Local<'a, v8::Value>, crate::Error>;
 }
 
-/// Allows all implementors of `serde::Serialize` to implement V8Serializable
-impl<T: serde::Serialize> V8Serializable for T {
+/// Allows all implementors of `serde::Serialize` to implement Serializable
+impl<T: serde::Serialize> Serializable for T {
   fn to_v8<'a>(
     &self,
     scope: &mut v8::HandleScope<'a>,
-  ) -> Result<v8::Local<'a, v8::Value>, serde_v8::Error> {
-    serde_v8::to_v8(scope, self)
+  ) -> Result<v8::Local<'a, v8::Value>, crate::Error> {
+    crate::to_v8(scope, self)
   }
 }