A more efficient replacement for `erased-serde` to box our async op returns.

It improves our op-baseline by ~60ns/op across all op-categories, which is a strong gain across the board, 20% of the current sync op baseline which is quite significant.

## Benches

```
Before:
test bench_op_async   ... bench:     772,851 ns/iter (+/- 55,538)
test bench_op_nop     ... bench:     265,945 ns/iter (+/- 13,634)
test bench_op_pi_bin  ... bench:     291,033 ns/iter (+/- 16,716)
test bench_op_pi_json ... bench:     285,994 ns/iter (+/- 7,377)

After:
test bench_op_async   ... bench:     710,543 ns/iter (+/- 28,573)
test bench_op_nop     ... bench:     210,362 ns/iter (+/- 3,354)
test bench_op_pi_bin  ... bench:     238,992 ns/iter (+/- 20,449)
test bench_op_pi_json ... bench:     223,458 ns/iter (+/- 3,186)
```

## Changes

- [x] Introduce `V8Serializable` trait
- [x] Drop `erased-serde` dependency
- [x] Maybe move `V8Serializable` to `serde_v8/`